blog2/requirements.txt
transformers>=4.21.0
torch>=1.12.0
pillow>=9.0.0
requests>=2.25.0
sentencepiece>=0.1.95
